Move ENV package to AWS
