+++ aws/env_test.go
@@ -1,4 +1,4 @@
-package env
+package aws
 
 import (
 	"os"
